Specify dependencies as an array of string package paths.
We want to be able to specify a dependency as something like the
following:

 - A depends on any version of B
 - A depends on B version 1.2.3 exactly
 - A depends on B version 1.2.3 or later

Specifying a dependency to a single `Package` won't cut it for that sort
of specification, so we're switching here to using package paths of the
form `user/package[/version]` - where `version` is omitted, the latest
package is used.

We don't yet do any validation of the contents of this array.