Create a rake task to download the packages
All the packages need to be download so they can create a local copy.
A rake command has been created for this purpose.

An existing `forge` server is required to act as the upstream source for
the existing package list.

The packages_url is then extracted and the zip files downloaded. The
downloads are done in parallel