Add a rake task to import packages
`Anvil` needs to import the packages so it can store the `metadata`
within its own db. This is accomplished using a `rake` task.

It uses the `ANVIL_LOCAL_DIR` env var from before, thus making it easier
to download and then import.

As `Anvil` does not host the content itself, it does not know the base
URL the content can be accessed from. It does however assume the
relative path from `ANVIL_LOCAL_DIR` is the same as `ANVIL_BASE_URL`.

Thus only the base url needs to be specified, in which the `package_url`
is inferred from.