Create a `packages:snapshot` rake task
This command runs the database setup, packages download, and import. It
will implicitly set the environment variables if they are not otherwise
set.

By default it will use the public forge server as the upstream, and
stores the packages in the `public/local-packages` directory.

If the `ANVIL_BASE_URL` is not specified, it will ask for the IP/ domain
instead. The protocol and `local-packages` are added automatically.

Highline is used to ask the question