Hack the dependency resolution during the snapshot
Anvil resolves dependencies when they are uploaded for the first time.
It does not allow a package to be added if it dependencies do not exist.

This is fine for the upload as the "developers" are, in essence, sorting
the packages. However this proves problematic for the snapshot which
is done alpha-numerically.

The snapshot does not import the packages through the same mechanism as
the upload, so the behaviour can vary. Previously it was skipping the
dependencies entirely on import. This meant the `forge-cli` client
(within Flight Direct) could not resolve dependencies. This bug has been
fixed in the Packages model.

However this also means the packages need to be imported in reverse
dependency order. The import has been hacked to repeat the import until
all the packages have been added. This retry mechanism assumes all
errors are due to missing dependencies and thus is may cause issues in
the future.

However it works for now