Detect when no more files can be imported
The import is complicated by dependencies as they are difficult to sort.
Instead they will be looped over until no more files are being saved.

This will either correspond to a successful import (in which all
packages are present) or a failure state.

There are numerous edge cases where a package can not be imported.
Currently not more reporting is being done, this could be improved