Create utility for generating zip files
In order to test the upload, zip files with different metafile.json
files are required. File fixtures are ill suited for this as it is only
the metafile.json that is changing. However the fixture will hide this
as they are zipped.

Instead creating the zip files dynamically makes it clearer what the
metadata file actually contains