Ensure the record can be updated without the zip
The zip file only needs to be validated when the record is created.
After this point, the zip file will not be available. To allow the
record to be updated, the zip checks are only done on create.

Most of the logic for this was already implemented. Only the finishing
touches (read: spec) was required.