Automatically assign packages to a default category
Previously this was done at the db level by hard coding the category
uuid. As this has been removed, it is now done before the package is
validated.

By using `find_by name`, it is more robust to the `uuid` changing (e.g.
moving the application somewhere else BUT not the db).